ecommerce-backend/requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
itsdangerous==2.2.0
Jinja2==3.1.6
Flask-CORS==4.0.1
gunicorn
MarkupSafe==3.0.3
psycopg2-binary==2.9.11
Flask-SQLAlchemy==3.1.1
python-dotenv==1.2.1
SQLAlchemy==2.0.44
typing_extensions==4.15.0
Werkzeug==3.1.4
